Full system hierarchy
I've added a top level, full system hierarchical schematic.

To make this work, I added sheet connections to all levels of the hierarchy, but so as not to upset anyone's current work, I saved the affected schematics to the top level (payload and avionics).  As a result this may not work because Avionics is a hierarchical schematic as well, and the links aren't preserved.

I've also brought in my most recent changes to comms to add the drivers  for the serial ports and to reflect the same signal naming convention (pin name_header-headerpin).